tp: refactor and simplify JSON tokenization
Over time, the ParseInternal function had grown into a bit of a monster
leading to a lot of subtle bugs. Fix this by refactoring it into
multiple functions which make tracking the execution flow a lot simpler.

Moreover, while we're here, fix a few bugs which cropped up as a result
of this around EOF handling.

Change-Id: Idc50a712d685181d4823f193191d3d726ecee3e7