[Fuchsia] Ensure that GetThreadId() is cached.
On Fuchsia a syscall is required to get ID of the current thread, so
ideally it should be cached. Previously GetThreadId() was trying to
cache it using thread_local variable, but it was inlined, so the value
wasn't actually cached. Moved the function to a new .cc file to a
function that's not inlined, which allows to cache the result properly.

Bug: 1166873
Change-Id: I2e3433856b393032af2a3a53847b5f42ddfe3872