Do not freeze or update interface if frozen: true
If an interface needs to be updated with .*-update-api, it can not be
marked frozen with `frozen: true`. It needs to be changed to `frozen:
false` before .*-update-api will work.
Same goes for .*freeze-api.

If a difference is detected when frozen: true, show an error message
that reflects that.

Test: make change to android.hardware.drm.IDrmFactory and build to see
new error message when `frozen: true`
Test: m android.hardware.drm-update-api (fails w/frozen, pass w/o frozen)
Bug: 231903487

Change-Id: If538f311da142156a21dae688423f35db40698ca