Implement FXT Embedded Perfetto Blobs
The Fuchsia tracing system is able to interface with Perfetto speaking
clients by wrapping Perfetto packets from them in a Blob Record header
and embedding them in a Fuchsia FXT format trace.

This patch adds support to the Perfetto Fuchsia trace importer for these
embedded Perfetto blobs. On seeing them, the importer writes them to a
secondary buffer which it forwards to the Perfetto Proto parser while
maintaining the existing context.

Doing this allows the simultaneous viewing of both events using native
Fuchsia tracing and Perfetto based clients running on Fuchsia in the
Perfetto viewer.

Change-Id: I3de9cc050f277a7fb2e35946f67068fc0c817e08