Add a rule for overriding waitForIdle() timeouts.
This works for CUJs on apps where UIAutomator can't find an idle state,
causing CUJs to take forever. See b/180944759 or linked bug for an exmaple.

It sets timeout in both androidx and android.support.test UIAUtomator
because they use separate timeouts and it's hard to know which version a
CUJ is using. Sometimes it's a mix.

Bug: 249550993
Test: Unit tests
Change-Id: Id42a30c05db10dfda24eec31ceff93f5eb6283c8