Reverse the order of the staging_dir_builder map
staging_dir_builder previously took a map of files to
their paths in the staging directory. If there were
duplicate staging directory paths, they were silently
ignored. This caused issues for mixed builds of flattened
apexes.

We could just add simple enforcement that there are no
duplicates, but it makes sense to reverse the order of
the map so that clients of staging_dir_builder can
quickly check for duplicates themselves and handle them
in a more appropriate context.

Bug: 254205429
Test: Presubmits
Change-Id: I5326e78783e3be51701f91dee376af7f9001e6eb