Fix crash when initialize fails.
If the initializing of the AndroidUnwinder object fails, some objects
will be left as nullptr. Unfortunately, the previous status of
the initialization call is not preserved, so a second unwind on
an improperly initialized object will crash. Add a variable to
keep track of the initialize status so that if it fails, all unwinds
from that point onward fail rather than crash.

Add a unit test to verify this works.

Bug: 256213419

Test: All unit tests pass.
Test: The new unit test fails on the old code and passes with the fix.
Change-Id: I1ef3383533cf00d8f41f7563208daac3d66a132b