Fix for print failure after printer certificate renewed
If the print client (DPS) detects a change in printer certificate,
it supposed to show a notification to the user to either accept
the new certificate and continue or reject. But as we found
recently, the BAD_CERTIFICATE status is getting treated as an
error reason rather than a blocked reason, and hence the job
is getting failed.

Test: Print through Wi-Fi discovered printer and then
reset the printer to factory default or change the ip
address of the printer and then print again
Signed-off-by: Saurabh Chauhan <mopriadevteam@gmail.com>

Change-Id: I056bde391c14999831710a7637d64ea953355272