Autoupdate of 2 submodules
*external/hyphenation-patterns updated to 4196330165c38e2031bb4add32d71f3543a3ca98
Update Georgian .chr.txt file

Previously it had each character in lowercase, repeated twice.
Replace the second occurance with the uppercase version of the
character.

This is needed because now that we've update mk_hyb_file.py to
python 3, it can correctly identify lowercase/uppercase Georgian
characters.

Bug: 203436762
Test: Presubmits
Change-Id: I39684439cd713b1146ef8318ed07b60c2effb355

*frameworks/minikin updated to f97b7b4a806bbbe056abc07c4c3750f9cfb7a62f
Convert mk_hyb_file to python 3

It already support python 3, except that some characters
are properly detected as lowercase in python 3 that weren't
in python 2. Add exceptions for 2 of them, and update one
of the .chr.txt files to make it work again.

Bug: 203436762
Test: Presubmits
Change-Id: I1f30bf717ab5ca915c7a0bf6acc87551ab6aea15