bazel apex: add support for generating _using.txt file.
This file contains NDK function symbols dynamically used by an APEX's
shared libraries.

The shared cc_stub_library should be linked with the --version-script
argument with the stub map file created by ndkstubgen, but isn't
currently done, so the generated stub so had unversioned symbols. This
CL passes the stub map into the linker args and additional linker inputs
for the cc_shared_library action.

Additionally, the using.txt generation script operates on a staging
directory, so this CL extends the staging dir builder action to also
return a directory (TreeArtifact) to be used as an input to the script.

Mixed builds integration to come in a follow up CL.

Test: new golden test
Test: analysis test
Test: presubmit
Bug: 239081455
Bug: 257226023
Fixes: 257226023
Change-Id: I5e83caa43ad50646db0e7d25eed5f2d75b47535a