script to visualize labels
I'm not sure a graphing program that allows you to manually
move nodes around - also the image cuts off some labels, so
if someone knows a better program or how to really use this,
that would be great. However, for now in the UI this presents,
you can move around the image to look at it.

(internal) demo'd here - b/251177105#comment34

Bug: 251177105
Test: cat example_dep_build_failure_output.txt | ./vis_aidl_ver_errors.py
Change-Id: I4baf430450e4a30fff4f712a55b7ce3e02d89e81