Add integration test for permission annotations
A new integration test is introduced with a server implementation which
relies on generated permission checks. This test is separated from the
other Java server as it is not yet compatible with the other backends
(i.e., C++, Rust).

Refactor the java_default rule for the java integration tests, so it may
be used for the new tests.

Bug: 241171714
Test: atest aidl_integration_test
Change-Id: Id68b71a073d64af1ec0bc3e2ccfbb9c694624b7f
Merged-In: Id68b71a073d64af1ec0bc3e2ccfbb9c694624b7f