Use block apex path as key for rootdigest overrides
When scanning block apexes in VM mode, apexd keeps rootdigest overrides
in a map for later verification. But using apex name as a key doesn't
work for "sharelibs" apexes because there might be two apexes for a
single sharelib apex name.

Instead, we can use block apex path as a key to a map.

Bug: 257377686
Test: ApexTestCases
Change-Id: Idd4e90f52e56b47d046c4eda88f34bf7b7aee4d5