p/android-4.14 - require CONFIG_NETFILTER_XT_MATCH_OWNER=y instead of QTAGUID
Technically this is not a valid change, as it restricts valid
configurations of the kernel, but... in practice... this is required
already by at least Android 11 (R) in order for traffic accounting
to work correctly.

When using xt_qtaguid Android R CTS fails:
  testTrafficStatsForLocalhost fails "java.lang.IllegalArgumentException: lower must be less than or equal to upper"
due to an extremely subtle difference in the implementation of
'-m owner --uid-owner XXX' (which xt_qtaguid breaks, by
effectively making the condition always true).

This was reported by Nvidia back in August 2021 in bug 196522401.

It takes a lot of painful debugging before we get to comment 36
which tries to make sense of things... and then comment 38 gets
to the root of the problem.

If someone does fail this (ie. VINTF) they'll need to flip:
  CONFIG_NETFILTER_XT_MATCH_QTAGUID -> n
  CONFIG_NETFILTER_XT_MATCH_OWNER -> y
  CONFIG_NET_CLS_BPF -> y

If someone actually *relies* on xt_qtaguid (and thus can't simply
flip the kernel config options) they'll need to request
a waiver for VINTF...  However, it's not at all clear how that could even happen considering the platform (userspace) side of the qtaguid implementation code was deleted in Android S (and was mainlined in T).
Additionally they'll then fail CTS, and we don't grant a waiver there.
So they simply won't be able to update to U (but I'm certain they
wouldn't have been able to run S or T either, so they'd need to be
upgrading straight from R).

Bug: 196522401
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I883e0c521e615a99f08361328dde2cd0d9a3d861