Fix faulty fragment handling in Storage
When performing any of the actions on the
Removable- or Device-disks inside of Settings->System->Storage,
the user was not returned properly to the storage summary list.
Instead, user was left in the storage detailed fragment with the
previous, now invalid, menu items still selectable.
Root cause was a missing adaptation to TwoPanel build,
so instead of moving a panel backwards as e.g "Eject" completed,
the backstack was popped as done in the original 1-panel.
This caused, not only strange ui behaviors, but also the side
effect that each of the fragments no longer could register
their focus change listeners, thus preview always remained
the same when stepping up and down in the list.

Fix also includes the removal of preview fragment as the
user has selected "Eject".

Bug: 258383118
Test: Go to Settings -> System -> Storage.
Plug removable USB storage device.
Highlight the removable storage and select "Eject".
Confirm that Left (main) panel is updated correctly,
while right becomes empty.
Press remote control key Back.
Confirm that navigation is made to "System" menu.
Press remote control key Up/Down in the list.
Confirm that preview panel is updated
for each focus change correctly.

Change-Id: I83ef9f55c62eb5e4ab52e4cabbb2e35279461b8c