Allow freeze-api to attempt to freeze a `frozen:true` interface
Interfaces aren't allowed to build with `frozen: true` when they have
changes between the last frozen and the current version, so we shouldn't
throw an error when we try to freeze an already frozen interface.
This allows us to make a change to the interface then freeze it directly
afterwards without the intermediate step of flipping `frozen: false`.

Test: freeze identity, then try to freeze it again (no-op)
Test: freeze identity, make a change, then try to freeze it again
(freeze next version)
Test: m android.hardware.identity-freeze-api
Bug: 258050166

Change-Id: I2cf6c615769ba893f700649f1fd7328be22dc485