Cause the lzma-sys crate to build and link liblzma.a
By default, the lzma-sys crate links against the system's liblzma.so.
This causes Cargo to propagate a dependency to the system library
directory to all of the libraries that depend on lzma-sys.

This CL sets an environment variable to cause llvm to build and link
against a static liblzma.a and also attempts to reign in pkg-config by
setting a specific PKG_CONFIG_PATH.

Test: ./build.py --lto thin --gc-sections --llvm-linkage static
Change-Id: Ie563737fce4cbff703505905d96576eafc207d68