Run test: Increase runner verbosity
Print stdout and stderr of all processes by default.
Remove the --verbose flag since that is the default behavior now.

Use ANDROID_LOG_TAGS=*:i by default, print all output to terminal,
but automatically filter it before we use it for stderr diffing.

Replace the --external-log-tags flag with --android-log-tags=*:v
which allows setting the log verbosity in one step.

Add --diff-min-log-tag which allows the tests to specify which logs
should be considered in stdout/stderr diffing (defaults to *:E).

Test: test.py -b -r --optimizing --64 --all-target
Test: test.py -b -r --host --run-test-option=--android-log-tags=*:v
Change-Id: Id70e68217a4b77e2764f517e775a07293fe99dcd