Add a no_std variant of libzeroize
The current "libzeroize" requires std support so add a new rust_library
that links against the core library, instead.

As the crate's "alloc" feature requires an allocator and as there isn't
a single standard implementation of it in no_std, don't enable the
feature for "libzeroize_nostd". Otherwise, given how Android builds Rust
libraries, this would force all users of the rust_library to use and
share the same allocator implementation.

Test: m pvmfw_bin # libzeroize_nostd linked in aosp/2280853
Change-Id: Ia971829111cd4678cedc16c7e8435a20476caa85