lift release checks to frozen interfaces
Now that we have the frozen bit, if an interface is frozen, we can
always do the release checks against it.

The error message right now, due to the layers in the release check
algorithm could be improve, but for now, it explains all the reasons
why the error may occur.

Fixes: 259145969
Test: build (w/ and w/o violation in tree + added tests)
Change-Id: I2a57a832a89a141a14ec8a7042c089e960280f5b