remove partially enforced unnecessary VINTF req
Previously we had a check that AIDL interface in the
VNDK must be stable, but we only did this when the
interface-wide (not per-backend) VNDK setting was set.
Also, this requirement doesn't matter, because the
runtime has careful checks. Also, VNDK libs could
provide vendor<->vendor unstable interfaces. Finally,
we have another process (VNDK current.txt) which
can get ahead of this problem.

Bug: N/A
Test: N/A
Change-Id: I4110f3184f982a1c46f75f970502ed58e53768c7