Autoupdate of 2 submodules
*packages/modules/Virtualization updated to f52ea1400ef26734c1f89f531ae350cc79d0f581
Merge "Run resigning Microdroid tests in a background process"
*tools/tradefederation/core updated to 039c61cddc3a36ae2d799432b7be4ccfd2276634
microdroid: Run VMs in a background process

Remove dependency on '--daemonize' by running the 'vm run-app' command
in a background process. This has the advantage that all VMs are
automatically killed when the test process dies for any reason, leaving
no resources used on the device.

Test: atest -p packages/modules/Virtualization:avf-presubmit
Change-Id: I2b60e463e61b3e6a18dad970f48842fed74c3630