WebView: Remove the ability to bisect WebView versions
This CL updates the WebView csuite tests to remove the bisection
that happens for test failures. In the test suite one WebView
version will be tested. The --channel or --webview-version-to-test
command line arguments can be used to install WebViews onto the
test device. If the test fails then the test will be rerun on
the preinstalled WebView. After that the test suite will exit
irregardless of the result.

Test: csuite-tradefed run webview-app-launch <params>
Bug: 206871132
Change-Id: I312e0df5a4ae848f935a80bbc33b883ae48cd8b2