Running Robotest class path correctly
Previously, Atest failed to resolve correct robolectric module name when
passing a file path. The original logic only works correctly for the
legacy robolectric, but will fail if it is the tradefed suppport one.

With this change, Atest is able to resolve the correct module name
without regarding the type of robolectric test.

Bug: 254379632
Test: atest-src packages/apps/Settings/tests/robotests/src/com/android/settings/uwb/UwbPreferenceControllerTest.java
Test: atest-src packages/apps/WallpaperPicker2/robolectric_tests/src/com/android/wallpaper/picker/BaseActivityTest.java

Change-Id: Ie812375edd6d58ca15f046a738eb43bd6ac098dc