tracing: add TraceBuffer::CloneReadOnly()
Allow to create read-only copies of a TraceBuffer. This is
a pre-requisite for non-destructive trace reads for bugreports.
The reason why the clone is read-only is because there is no
plan to keep writing into these snapshots. On one side, I think
everything would work even if we wanted to write onto it.
OTOH I don't want to create the illusion that this will work
given that the upcoming CLs will use this only in read-only mode.

Test: perfetto_unittests --gtest_filter=TraceBufferTest.Clone_*
Bug: 260112703
Change-Id: I355515e97519396f492b9b15fa769adef02cf731