Remove current versioned snapshot
Versioned snapshots are no longer generated for, or used by, Soong
which means that having a versioned snapshot in an Android.bp file
processed by Soong is just a waste of resources and prevents all the
support for versioned snapshots from being removed from Soong.

This change removes the @current versioned instances of the modules
from the sdk snapshots.

Merged-In specifies the id of a downstream change that updated the
module sdk snapshot with one which does not contain @current
versioned modules and so does not need to have this change applied.

Bug: 260237150
Test: m nothing
Merged-In: I68186557189c35a943c78e4195e395960a8ef8c6
Change-Id: I4cc51c5a2d9eefbeb1feaeb8242105e85f5e473f