Remove current versioned snapshot
Versioned snapshots are no longer generated for, or used by, Soong
which means that having a versioned snapshot in an Android.bp file
processed by Soong is just a waste of resources and prevents all the
support for versioned snapshots from being removed from Soong.

This change removes the @current versioned instances of the modules
from the sdk snapshots.

Merged-In specifies the id of a downstream change that updated the
module sdk snapshot with one which does not contain @current
versioned modules and so does not need to have this change applied.

Bug: 260237150
Test: m nothing
Merged-In: I1c7ee7882079312a8a10e4ee2509dd424080b6be
Change-Id: Ib9e1ec144104e072222bb3d4f7bd77d970cfb456