Update domination tree when removing tries, if the catch doesn't exit
When eliminating the unnecessary try and its catch block we turn the
TryBoundary instructions into Goto instructions. If one of these
instructions is pointing to the exit block, we use its single
predecessor instead. If this TryBoundary-turned-into-Goto instruction
was the only one pointing to the Exit, we also have to update the dominators.

Bug: 260387991
Fixes: 260387991
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Test: dex2oat compiling the app in the bug
Change-Id: If05b60e81b2441c4c2235ffad7649813b98d3d69