vndk_use_version required for unfrozen vndk iface
Because, when freezing, the version in the VNDK would
change, and we want freezing to be automatable.

Note - this requires vndk_use_version even when there
is only one version available, and in this case, we
already have code to make sure the lastest version
(which is unfrozen) is not used in the dessert branch.
However, that other code is a hack, and I do this
in preparation of being able to remove the hack and
have even more explicit configuration.

Fixes: 259465531
Test: build (aidl_test)
Change-Id: I999a8626e5bed81125bac0bc04d3e3856adc6cf3