Add a test config template for doh_ffi_test
resolv_test_config_template is used for the tests that require root
access. We can't use it for doh_ffi_test because doh_ffi_test needs
to be run by mts-user that runs tests on user build.
So add a test config template that doesn't contain RootTargetPreparer
for doh_ffi_test.

Bug: 259349997
Test: run mts-dnsresolver -m doh_ffi_test on a user build
Change-Id: I0c4f791ead2fd3ec16b2b94ef5403e1b73254c11