Autoupdate of 2 submodules
*prebuilts/sdk updated to 0db080fbf11c20e15d1c6fc1c304afb9bc9c1eaf
Merge "Adding missed incompatibilities due to metalava bug"
*tools/metalava updated to f6f476237f3af2b946953783dc320b45c10c04a2
Metalava fails to detect incompatibility in module build

Incompatible changes to a released System API were not detected, this CL fixes the problem caused when a partial codebase is passed to MergedCodebase and merged with a full codebase removing the changes made in the partial codebase. When the list codebases (mergedcodebase) is passed to ComparisonVisitor.Compare() the first one is  used as master so the new codebase should be first or the changes are lost.

Bug: 223771856
Test: ./gradlew test
Change-Id: Ic314b5c220074339f0f7f2df26cf337b2dc1d6de