Add bpttool to the trout source tree
This is a modernized (Python 3 compatible) version of the older bpttool
used for some IoT projects within Android. The tool was since removed and
Cuttlefish uses a different incantation to produce disk images. However, the
Cuttlefish tool is highly specific to crosvm

bpttool is a simpler tool whose job is simply to create a "disk" from a set
of partitions file, i.e. add a valid GPT and concatenate the partition files
into a file that can be used a standalone virtual disk and flashed directly
to a partition or a host file and used to launch an Android GVM as a
virtual guest disk

Since this tool can be used by trout, we are bringing it back and
supporting it as a way to craft a trout image to be launched under
any automotive hypervisor

Bug: 260658446
Bug: 235196371
Test: the trout team uses this tool internally
Change-Id: I4cfd595d6b68740feec86bb4371d32effad367d9