Allow for provisioned keys to be used in tests
Some of the tests were provisioning a full key pool (20 keys as of now),
and that might take a little bit of time. During that window, we started
to see a keystore client consuming one of the keys. This breaks our
assumption that nobody could possibly have consumed a key.

Change the tests to instead do greater than compares instead of
equality.

Bug: 260874294
Test: RemoteProvisionerUnitTests
Change-Id: If3a7be6d92500db8bc5a0443dab6c27a12cb03b0