Distinguish SD secrets for key upgrade
Fix a misunderstanding in the existing code: a key upgrade operation
should *not* remove the secure deletion slot for the existing keyblob;
instead, Keystore will subsequently issue a separate `deleteKey()`
operation for the old keyblob.  (This add-new then delete-old behaviour
is more robust if there is a failure between the two operations.)

Map this through to the `SecureDeletionSecretManager` trait, in the form
of an additional parameter on the `new_secret` method.  The expectation
here is that implementations should reserve some extra space for these
upgraded keys (under the expectation that this space will be freed once
the whole process completes), to make OS upgrade procedures more
reliable.

Bug: 197891150
Bug: 253948023
Test: VtsAidlKeyMintTargetTest
Change-Id: I0fe95a1e885ddeb11b556cd3eec7874982e2dd8b