Add JDWP metrics to DDM traffic
Record processing start and duration for each cmd. Flush the content of
the timing buffer when it is full.

At the end of the session, (identified by VM::Exit cmd), also flush.

The timing data is sent over a DDM packet with all
the metrics recorded to allow debugger performance profiling.

Test: NA
Bug: NA
Change-Id: If988dc25a8f7e552f1b049dbed8a953e47a7c6d0