api-versions.xml: emit `sdks="0:next_dessert_int"` for non-finalized APIs
A method or field that has not been finalized (either as part of an
Android dessert or an SDK extension) needs an explicit `sdks` attribute
to prevent it from inheriting an incorrect value from its parent
(finalized) class.

One option to indicate non-finalized APIs would be to use an empty list
(sdks=""). Another option would be to re-use the value of the since
attribute (sdks="0:next_dessert_int").

Teach metalava to go with the second option as this allows Lint to
determine that the symbol will be available in a future version as
opposed to treating it as a non-existing symbol.

Bug: 260515648
Bug: 260716454
Test: ./gradlew test --tests ApiToExtensionsMapTest --tests ApiGeneratorTest
Change-Id: I564454494c07d47f03dfa0254eabfd51cbdb9668