Change tables to use colgroup; delegate layout to first row
Also adds a div(classes = "list") around lists of SymbolDetails.

Delegating layout to the first row has precedent: it's how HTML itself
works.
Because of the type parameter on the table/classlike, we know all rows
must be the same type, and therefore want the same layout.
We can't use the type param for layout, though, b/c type erasure.

I find this an ugly solution, but I'm not sure what would be better.
Definitely open to ideas.

(this test forked from KMP changes because it touches non-KMP results)

Test: changes to all integration tests
Change-Id: I85c9182ab56ab642e20367162f1ca05e85b9551d