Rewrite proto_logging Android.bp to compatible with bazel.
When converting from soong to bazel we use filegroup to represent
external proto source files, in which case the filegroup will be
converted to bazel proto_library. This requires the filegroup to be self
contained, i.e., all the imported proto files have to be included in the
filegroup, otherwise we have to reference them as srcs.

Bug: 261237265
Test: Manual test
Change-Id: I67ee64abd605d8d3cc7d19bc4542e941c98ea857