Run test: Execute scripts via runner bash script.
Simplify all run tests to simple execution pattern:
adb push <inputs> && adb shell <runner> && adb pull <outputs>

This dramatically reduces the number of adb commands, and
it will make it trivial to reimplement the above steps in tradefed.

Follow the above logic for host and jvm as well for consistency.

Test: ./art/test.py -r --all-target --optimizing --64
Change-Id: I4b54a4d32195c440ca43133bbbe988ae24e626af