Autoupdate of 2 submodules
*frameworks/base updated to 17b0dccd447f8e898a8edd4f9c821fb648f603af
Merge "Allow non-updatable modules to contribute to @TestApi"
*packages/modules/Virtualization updated to 147eb6f93095f1fe6778f5cdd843ddf103fb52cf
Move some VM config APIs to @TestApi

Another change in this topic allows java_sdk_libraries belonging to
non-updatable APEXes to contribute to test API stubs.

Since  getPayloadConfigPath() and setPayloadConfigPath() are not
expected to be part of the @SystemApi, move them to @TestApi.

Bug: 261004711
Test: presubmit
Change-Id: I48a19ed925cc0c75cf6a965699aaa291508ea7d7