Try to fix art-linux-bionic-zipapex
Use more robust version of `ln` command.

Some tests will run themselves twice and thus run `ln` twice,
which results in creating link inside the link (it should be
harmless, but it somehow seems to cause an error sometimes).
I was unable to reproduce the error locally.

Use -T with `ln` to make it explicit the second argument
is the destination, not a directory of the destination.
I believe this should potentially help.

Test: run_build_test_target.py art-linux-bionic-x64-zipapex
Change-Id: I2c263d9fa17702b545732a446ff00ff128ffc413