Autoupdate of 2 submodules
*build/soong updated to da62908229eaaa7c4521bfed1c815a8c2eb8173a
Merge "Vendor apex sets "vndkVersion" when it uses vndk libs"
*system/apex updated to 774a236d26beef4893cf98950eb053f3038579ed
Reject if vndkVersion mismatches for vendor apexes

Vendor APEXes can rely on VNDK libs on system. When a vendor apex is
installed on a device with a different VNDK version, native modules in
the apex will fail to run. (Note that VNDK libs are not stable across
versions)

We'd better reject such cases at install/stage time.

In this change, apexd rejects APEXes when vndkVersion is set in
apex_manifest and the version mismatches with the actual VNDK version of
device.

Bug: 222620439
Test: VendorApexHostTestCases
Change-Id: I34126c9e4a22d4a3ae327ffbe914bbce3babc585