Autoupdate of 2 submodules
*external/adt-infra updated to e6b544b9387041c40fa501dd5afdce84f7e3ab8b
Only detect running emulators

In our test case the discovery files we write are usually the same
between various runs. The only thing that changes is the PID.
It is possible that the emulator gets forcefully terminated and does not
get a chance to clean up the discovery files.

Previously we could discover emulator discovery files that had no
matching pid.

The emulator discovery api has improved and now:

- Only returns running emulators
- Can properly terminate an emulator

This improvement introduces a new dependency (psutil) which has been
added.

Bug: 258276008
Change-Id: Iee17cd51b250aa9e9e7a900db57e95cabd0b1cc5

*external/qemu updated to 04a759662e863c1dc124253ba2ac54364769787f
Merge "Only discover emulators that are alive" into emu-master-dev