rkpdapp unit tests no longer instrument the app
RkpdAppUnitTests were targeting the com.android.rkpdapp package.
However, this package is named com.google.android.rkpdapp in shipping
configurations. Additionally, this renamed package is signed with a
Google-owned key.

Therefore, it's not reliable to target the AOSP package name on all
platforms. To work around this, we rebuild all the app sources into a
test library so that we can unit test the classes used by the app.

This change required a few test fix-ups due to:
1. We are not inheriting permissions from rkpdapp
2. The test is using SDK, not platform apis

Bug: 261862422
Test: atest RkpdAppUnitTests
Change-Id: I56874e5d8558613ddcd9a9f9a0cd8585b00b0d11