Extend baseline to support api level filter
The tradefed applies to all branches from S. However, the target of a new baseline is always the newest android version because tests assuming a wrong baseline may need to be modified and we don't want to change the test that have alreadly been released. Therefore, we extend the baseline so that each baseline can be filtered by the api level.

Test: presubmit test
Test: manually
Bug: 260657484
Change-Id: I76fe2d4d653345d415fa7b99ab5dcf2b5d564085