Test: Deflake RepeatedSetup_KeepChangingPrivateDnsServers
The test fails if the flag dot_async_handshake is 1. If the
flag is 0, DnsResolver doesn't retry when DoT connection
establishement fails; if the flag is 1, DnsResolver does retry
when DoT connection establishment fails. To fix the test to not
fail when the flag is 1, we check whether |connectCountsAfter|
is greater than |connectCountsBefore| rather than check the exact
value returned by tls.acceptConnectionsCount().

Bug: 261327438
Test: run the test with dot_async_handshake=1
Test: run the test with dot_async_handshake=0
Change-Id: I729e5a91d97bd57dbbe5670de2305bc7f8a09200