Consolidate erofs/ext4 fstabs
Cuttlefish will soon use erofs, but we have a long term desire to
support ext4 for GSI testing, so for now we must have both fstab
entries. erofs is tried first to reduce errors in the normal path, but
since we haven't converted the default this yet this is testing the
fallback works (which it does).

Bug: 202022157
Change-Id: Ia7b84d0c3c2ea1fc0f757d21e7092e3b1d7ec816