Autoupdate of 5 submodules
*bionic updated to 137ecdd059e15af551ef258d21c59f4dce6086f6
Merge "Update to v6.1 kernel headers."
*external/kernel-headers updated to 10912b7c6aa15a6763907f45f3f39260c295b486
Merge "Update to v6.1 kernel headers."
*external/perfetto updated to 3f25b2b1fa788574cbdddce54d1df1dc3bbcb4d3
Merge "Update for kernel 6.1 headers."
*external/rmi4utils updated to acae70240d5d8f4b662cdf450dd395c19d91999d
Fix code to compile with new kernel 6.1 headers.

The new kernel headers conflict with some of the names used in
the hiddevice.cpp file. Change all of those so they are unique
to that file and will not conflict in the future.

Test: Builds.
Test: Boots on panther.
Change-Id: I46010a4c3b1914fa0c51ec840663cb75aadfd6a2

*system/extras updated to 591609c755640ae45e4a7d1180a429c2aee6a3d5
Merge "Update for kernel 6.1 headers."