Disable test 305-other-fault-handler in switch interpreter mode.
This test intentionally segfaults and expects the fault to be handled,
but the fault handler in ART is only installed if implicit null / so /
suspend checks are supported, which is not true in switch interpreter
mode.

Prevously the test didn't fail on interpreter, as the implicit checks
are enabled for all supported architectures.

Bug: none
Test: art/test.py --target -r --64 --interpreter 305-other-fault-handler
      # observe that the test is skipped
Test: art/test.py --target -r --64 --optimizing 305-other-fault-handler
      # observe that the test runs successfully
Change-Id: I4d17e992e24d29fd8ea8525ec95050541d8207c8