Use shared libc++ on device
The runtime apex already has artifacts depending on libc++.so. So, on
device, using shared libc++.so can save us some storage.
(On aosp_cf_x86_64_phone-userdebug, it's reduced 662944->450632, which
affects the sizes of runtime and virt apexes.)

However, on host, linkerconfig is statically linked with libc++ for
convenience. (No extra setup needed to run linkerconfig on host for
libc++.so).

Bug: 262330207
Test: m && boot
Change-Id: I47b9192cfe6142f327a4eb1a160b751a887420d6