minijail: add local definitions of MADV_FREE and __NR_membarrier
MADV_FREE and __NR_membarrier are used internally by musl libc and
therefore added to the crosvm seccomp policies by b/258117555, but
are not defined by the old glibc prebuilts in the Android platform.
Add defines for MADV_FREE and __NR_mmebarrier to gen_constants-inl.h
and gen_syscalls-inl.h.

Bug: 263152407
Test: launch_cvd --cpus=4 --memory_mb=4096 --console=true --enable_sandbox=true --base_instance_num=1 --report_anonymous_usage_stats=n --undefok=base_instance_num,console,enable_sandbox,report_anonymous_usage_stats
Change-Id: I9633e7403f924cdf454c9dc3bebae7740f621091