Embed libbssl_rust_support within libbssl_ffi
Include the whole of the libbssl_rust_support archive in libbssl_ffi
rather than just referencing the archive to link against. This is safe
because libbssl_rust_support is not a public targe, so there won't be
double linking, and it helps resolve linking errors when the static
version of libbssl_ffi turns up in the dependency tree.

Test: TH
Change-Id: I3cfbfaba7124da21bc2124ac7f9cf43b5eedc2f1