Autoupdate of 2 submodules
*art updated to cc77cdec4eeb4299efa1a94d942cec99ffe16adb
Remove duplicates in ART apex

bootclasspath_fragment causes the dependent libs to be in the apex.
Having those libs in java_libs as well is confusing because
bootclasspath_fragment might do something with those deps but it's not
clear which artifacts are used in the apex.

Now, it's clear that those dependencies are from bootclasspath_fragment,
not from java_libs.

Bug: 191770320
Test: m com.android.art
Change-Id: I17df5fc25c1d518e38155509a8b111545ec4305a

*build/soong updated to f7562aee8b36d05293e4d367314f7569d092eeca
Merge changes from topic "dup-check-in-apex"

* changes:
  apex_vndk doen't need stub variants
  Check for duplicate bundled files from different modules.