Make PublicKey a simple wrapper of PKey
Remove the double representation of the key and only keep the PKey. A
PublicKey will only be created with one of the supported kinds of key
represented by the Kind enum.

Encode to PEM for printing rather than raw hex strings as it's more
interoperable with other tools.

Add tests that try creating PublicKeys from different kinds of PKey.

Test: atest libcert_request_validator_tests
Change-Id: I04b1a60a3857c331e813ccad90bf3985c00d328c