Don't instrument stack with instrumentation exit PCs
In preperation to remove instrumentation stubs don't use instrumentation
exit pcs to call method exit hooks for methods already on stack. For
runtimes that started off as debuggable we should never need this. For
non-debuggable runtimes the support is best effort and it is OK to miss
few method exit hooks for methods on stack.

Bug: 206029744
Test: art/test.py

Change-Id: I8e8a587c7f7d249a5bffea8309d7c25c30639355