Do not set failing local cuttlefish instances to be in use
This commit changes the conditions for acloud to modify the lock files
of local cuttlefish instances. `acloud create` sets the instances'
state to in-use only if it succeeds.

This commit adds --local-instance-dir as the other condition. It is
needed for the old version of the testing infrastructure that relies on
acloud to choose instance IDs. The new version supports allocating
instance IDs, so the condition will be removed after the new version is
deployed.

Test: acloud-dev create --local-image --local-instance -vv
Bug: 261109137
Change-Id: Ie0bc33804a18a9c4079f85139537790fd7aff841