Delete mystery tests
The origin _CBOR_* files doesn't seem to be known and we don't have the
root key so get rid of them. These tests have been getting in the way of
refactoring because they deal with fragments of a chain that don't have
a root, but this isn't behaviour we need to support.

Also remove the unused device info test data while we're deleting
things.

Test: atest libcert_request_validator_tests
Change-Id: Ice37e010b8f80fe3cbe6e73ca5f12f091e8fb60f