Revert the changes to test/2246-trace-stream
Partial revert of aosp/2361086 to fix failures on the bots. The output
of method tracing is not consistent across all configurations for
non-debuggable runtimes. For ex: optimizing has fewer entries due to the
use of more oat code but no-image config has more entries. Revert the
changes for now. In a follow up CL we could only check for the classes
introduced by the test class to have consistent output.

Test: art/test.py -t 2246
Bug: 259258187
Change-Id: I2e0094015123031751993c0f18a35ec33a269da9