Do not rely on symlink-prefix
_copy_api_contributions used a symlink-prefix hack in Bazel to create
the convenience symlinks in an output directory, and then copy the files
via that symlink path. However, the symlinks were not being created in
my latest workspace (possibly after the recent bazel update?) and
therefore API export started to fail. `--symlink_prefix` is expected to be deprecated
soon (https://bazel.build/reference/command-line-reference).

The fix here is to not rely on symlinks and query for the
output_user_root using `bazel info`. It adds one more bazel call, but it
should not increase the overall build time significantly.

This CL does not remove `symlink_prefix` altogether to prevent the logs being
spammed with the warning "cannot create bazel symlink"

Test: test_build.sh

Change-Id: I5fee75c738f219a72859d777e51568788e4e7225