Add a provider to supply adb login arguments using a login files dir.
AppCrawlTester currently needs to be explicitly supplied with either a Roboscript or CrawlGuidance file through options.
This change adds a provider which finds either the Roboscript or Crawlguidance file for the given app package and adds the adb arguments to pass to the crawler.
This still requires passing the directory with login files to AppCrawlTester.

Test: atest csuite-harness-tests

Change-Id: Ibe434e57c5ccaece05cc105fd6847ffb0cb935c1