Autoupdate of 2 submodules
*external/rust/crates/serde updated to 1fbe19a54c5d251ff0eff9b3c5cd4ce0bf093cee
serde: Enable the “rc” feature

This allows us to serialize and deserialize fields containing “Rc” and
“Arc” types. I need this for testing PDL, but it could potentially be
useful for others as well.

Note that enabling the feature simply allows Serde to generate code in
cases where it would fail with an error before.

Test: presubmit
Change-Id: Ie39d6b842c1718f0f2c4f824388a324df271804e

*packages/modules/Bluetooth updated to cae1216893b72da3cd8747cfa8361769a07fbe74
pdl: Test canonical vectors using Serde

Being able to serialize data with Serde allows us to test enums using
the canonical test vectors: we can take the output of PDL and
serialize it to JSON, which we then compare with the canonical test
vectors.

While this was motivated by the need for testing, Serde support can be
useful in general so I think this is an overall nice change. The
support is behind an off-by-default feature flag, just in case people
don’t want the generated code.

Test: atest pdl_tests pdl_rust_generator_tests_{le,be}
Change-Id: I6a19a939df7b449d350dbcda8a7daa2cb96733a2