Don't turn off method exit hooks if a frame needs force deopt
Check if there is any frame on the stack that requires a force deopt
before turning off method entry / exit hooks. JITed frames are marked
for a force deopt after structurally redefining any method. Since the
offsets of the fields are embedded in the JITed code it's not safe to
use the code. As long as some frame on the stack requires a force deopt
we enable method exit hooks.

Bug: 206029744
Test: art/test.py
Change-Id: Iad9a6c79f96f41a55ab1d24a52662d37fc8e7e64