Don't link libcxx against the gwp-asan crash handler.
This is unnecessary, as GWP-ASan's crash handler specifically avoids
using libcxx.

This allows us to link the crash handler into the linker, in future, to
support the recoverable mode.

Also, don't build a DSO variant of this. It shouldn't be used by anyone,
and linking causes a cyclical dependency on libc symbols.

Test: build android
Bug: N/A
Change-Id: I5c844ce2f4d610ed52674fe4d6c69e9781f90e98