Extend flags passed to Trace::Start to also specify the clock source
We should enable users of method tracing to choose the clock source.
Earlier the clock source could only be chosen at the boot time and
wasn't very useful. This CL extends the flags that's already used by the
method profilers to specify the clock source. The clock source could be
wall clock, thread cpu or both. Thread cpu is an expensive operation and
hence we should enable it only when requested.

Drive-by fix: Also expose interval from Trace, so we can use it in the
post zygote hooks.

Bug: 259258187
Test: art/test.py

Change-Id: I6850fe2dc90d9d1e7f9408b5fd4bfbf390bc7ea3