Create record_binder tool
record_binder is a tool for starting and stopping the recording of
binder transactions for a service.

record_binder is the same aidl analyzer binary with no analyzers. When
no analyzers are present, the list command is removed. Without
analyzers, thus no definition of specific transactions, record_binder
cannot inspect transactions in detail. All the rest of the functionality
probably still makes sense for record_binder, including the planned
support for replaying recorded transactions.

Test: Make a fresh image. Use and explore with `record_binder`
Change-Id: Ie678419d63e5d7777ee77826fcf3adba6875dcb9