Add a Cargo build option
The Android build is often slow and requires setting up the whole repo
to build anything so it can be convenient to use Cargo for a faster
development loop. Cargo will be a secondary build option that isn't
checked by CI so its maintenance is "best effort".

Only one small tweak is needed to the integration tests in order to make
Cargo work seamlessly with the current state of the project.

Test: atest hwtrust_cli_tests
Test: `cargo test` with a Cargo.toml manually derived from Android.bp
Change-Id: Ic114a9b96e84d1abf5086a31c5bbca9cd9e58853