Delete kernel_fcm_version >= target_fcm_version requirement.
The kernel branch (KMI) in the kernel release string
is not meant to be compared with VINTF target FCM version
(manifest levels). Hence, this requirement is deleted.

Note: The GkiComplianceTest.GkiComplianceV2* tests will continue
to ensure that the kernel_fcm_version is specified. In addition,
kernel config checks requires the kernel_fcm_version to be specified.

Bug: 255995840
Test: TH

Change-Id: I80b418bc01ef0c8b30cd7441f92fb18bde7c4452