ANDROID: Fix riscv64 startup with LLVM toolchain
The value of "register gd_t *gd asm ("gp")" is restored to zero after
assignment with the LLVM toolchain. This seems similar to an issue
noted in the U-Boot code for ARM64, where gp is set up in assembler,
but to keep things simple use some inline assembler and implement
the arch_setup_gd() hook like what is done for x86.

Bug: 254280358
Signed-off-by: Alistair Delva <adelva@google.com>
Suggested-by: 毛晗(猛汉) <maohan.mh@alibaba-inc.com>
Change-Id: Ia1ebd2954c2fc030b6e3d5c5f89d49bc188c729f