tp: annotated callstacks: heuristic for the "switch" interpreter implementation
Quoting David: the "switch" interpreter is a fallback for configurations
that "nterp" does not support, which should be rare (e.g. debugging).

Name mangling explanation (_ZN3art11interpreter = art::interpreter::)
 _Z  - mangled prefix
 N   - namespace (technically, nested-name in ABI docs)
 3   - name of length 3
 art - literal
 11  - name of length 11
 interpreter - literal

This works even for templated types (which have return types encoded in
the mangled name), as the return type is towards the end of the mangled
string.

Bug: 191856355
Change-Id: I7b216ac8769855392f6301700e208e766f3d44d8